==> notebook2/1_data_list (1).ipynb <==
import sys
import os
import gc
sys.path.append(r"..")

from utils import *
import pandas as pd
import numpy as np

path_build = "./build/"
path_embed = "./embed/"
path_list = "./feature_series/"
path_mnt = "./mnt/"
path_sub = "./sub/"


pd.set_option('display.max_columns', None)

click_log = pd.read_pickle(path_build + 'click_log.pkl')
train_user = pd.read_pickle(path_build + 'train_user.pkl')
print(click_log.shape, train_user.shape)

click_log.sort_values(by=['user_id','time'], ascending=[True,True], inplace=True)
click_log = click_log.reset_index(drop=True)
click_log = click_log[click_log['click_times']<=8] #去除点击次数99.99%外的数据记录
print(click_log.shape)
click_log.head()

# click_log['creative_id_times'] = click_log['creative_id'].astype(np.int64)*10000000 + click_log['click_times']
# click_log['ad_id_times'] = click_log['ad_id'].astype(np.int64)*10000000 + click_log['click_times']
# click_log['product_id_times'] = click_log['product_id'].astype(np.int64)*10000000 + click_log['click_times']
# click_log['advertiser_id_times'] = click_log['advertiser_id'].astype(np.int64)*10000000 + click_log['click_times']
# click_log['product_category_times'] = click_log['product_category'].astype(np.int64)*10000000 + click_log['click_times']
# click_log['industry_times'] = click_log['industry'].astype(np.int64)*10000000 + click_log['click_times']


# click_log['creative_id_times'] = click_log['creative_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
# click_log['ad_id_times'] = click_log['ad_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
# click_log['product_id_times'] = click_log['product_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
# click_log['advertiser_id_times'] = click_log['advertiser_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
# click_log['product_category_times'] = click_log['product_category'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
# click_log['industry_times'] = click_log['industry'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')

click_log['product_advertiser'] = click_log['product_id'].astype(np.int64)*10000000 + click_log['advertiser_id']
click_log['product_industry'] = click_log['product_id'].astype(np.int64)*10000000 + click_log['industry']
click_log['advertiser_industry'] = click_log['advertiser_id'].astype(np.int64)*10000000 + click_log['industry']

click_log.head()

click_log = multi_column_LabelEncoder(click_log, 
                                      columns=['product_advertiser','product_industry','advertiser_industry',
                                              ])

print(click_log['product_advertiser'].nunique())
print(click_log['product_industry'].nunique())
print(click_log['advertiser_industry'].nunique())
# print(click_log['product_category'].nunique())
# print(click_log['advertiser_id'].nunique())
# print(click_log['industry'].nunique())
# print(click_log['creative_id_times'].nunique())
# print(click_log['ad_id_times'].nunique())
# print(click_log['product_id_times'].nunique())
# print(click_log['advertiser_id_times'].nunique())
# print(click_log['product_category_times'].nunique())
# print(click_log['industry_times'].nunique())

from collections import defaultdict
import sys
# creative_id = defaultdict(list)
# ad_id = defaultdict(list)
# product_id = defaultdict(list)
# product_category = defaultdict(list)
# advertiser_id = defaultdict(list)
# industry = defaultdict(list)
# click_times = defaultdict(list)
# time = defaultdict(list)
# creative_id_times = defaultdict(list)
# ad_id_times = defaultdict(list)
# product_id_times = defaultdict(list)
# advertiser_id_times = defaultdict(list)
# product_category_times = defaultdict(list)
# industry_times = defaultdict(list)

product_advertiser = defaultdict(list)
product_industry = defaultdict(list)
advertiser_industry = defaultdict(list)


for user,f1,f2,f3  in zip(
                             click_log['user_id'], 
#                             click_log['creative_id'],
#                             click_log['ad_id'],
#                             click_log['product_id'],
#                             click_log['product_category'],
#                             click_log['advertiser_id'],
#                             click_log['industry'],
#                             click_log['click_times'],
#                             click_log['time'],
#                             click_log['creative_id_times'],
#                             click_log['ad_id_times'],
#                             click_log['product_id_times'],
#                             click_log['advertiser_id_times'],
#                             click_log['product_category_times'],
#                             click_log['industry_times'],
                            click_log['product_advertiser'], 
                            click_log['product_industry'],
                            click_log['advertiser_industry']
):
#     creative_id[user].append(f1)
#     ad_id[user].append(f2)
#     product_id[user].append(f3)
#     product_category[user].append(f4)
#     advertiser_id[user].append(f5)
#     industry[user].append(f6)
#     click_times[user].append(f7)
#     time[user].append(f8)
    
#     creative_id_times[user].append(f9)
#     ad_id_times[user].append(f10)
#     product_id_times[user].append(f11)
#     advertiser_id_times[user].append(f12)
#     product_category_times[user].append(f13)
#     industry_times[user].append(f14)
    
    product_advertiser[user].append(f1)
    product_industry[user].append(f2)
    advertiser_industry[user].append(f3)
user_ids = pd.DataFrame({
                        'user_id':list(product_advertiser.keys()),
#                          'creative_id':list(creative_id.values()),
#                         'ad_id':list(ad_id.values()),
#                          'product_id':list(product_id.values()),
#                          'product_category':list(product_category.values()),
#                          'advertiser_id':list(advertiser_id.values()),
#                          'industry':list(industry.values()),
#                          'click_times':list(click_times.values()),
#                          'time':list(time.values()),
                         
#                          'creative_id_times':list(creative_id_times.values()),
#                          'ad_id_times':list(ad_id_times.values()),
#                          'product_id_times':list(product_id_times.values()),
#                          'advertiser_id_times':list(advertiser_id_times.values()),
#                          'product_category_times':list(product_category_times.values()),
#                          'industry_times':list(industry_times.values()),
                        
                        'product_advertiser':list(product_advertiser.values()),
                        'product_industry':list(product_industry.values()),
                         'advertiser_industry':list(advertiser_industry.values()),
                        })
print(user_ids.shape)
user_ids.head()

path_list = "./feature_series/"
for name in ['product_advertiser', 'product_industry', 'advertiser_industry']:
    np.save(f"{path_list}{name}_list_int.npy", user_ids[name])
    user_ids[name] = user_ids[name].map(lambda x: [str(i) for i in x])
    np.save(f"{path_list}{name}_list.npy", user_ids[name])
