==> notebook2/1_data_list.ipynb <==
import sys
import os
import gc
sys.path.append(r"..")

from utils import *
import pandas as pd
import numpy as np

path_build = "../../data/tencent2020/build2/"
path_save = "../../data/tencent2020/save/"

pickle_path = "../../data/tencent2020/pickle"
if not os.path.exists(pickle_path):
    os.mkdir(pickle_path) 
pd.set_option('display.max_columns', None)

click_log = pd.read_pickle(path_build + 'click_log.pkl')
train_user = pd.read_pickle(path_build + 'train_user.pkl')
print(click_log.shape, train_user.shape)

click_log.sort_values(by=['user_id','time'], ascending=[True,True], inplace=True)
click_log = click_log.reset_index(drop=True)
click_log = click_log[click_log['click_times']<=8] #去除点击次数99.99%外的数据记录
print(click_log.shape)
click_log.head()

click_log['ad_id'].min(), click_log['ad_id'].max(), click_log['ad_id'].nunique()

click_log['creative_id_times'] = click_log['creative_id'].astype(np.int64)*10000000 + click_log['click_times']
click_log['ad_id_times'] = click_log['ad_id'].astype(np.int64)*10000000 + click_log['click_times']
click_log['product_id_times'] = click_log['product_id'].astype(np.int64)*10000000 + click_log['click_times']
click_log['advertiser_id_times'] = click_log['advertiser_id'].astype(np.int64)*10000000 + click_log['click_times']
click_log['product_category_times'] = click_log['product_category'].astype(np.int64)*10000000 + click_log['click_times']
click_log['industry_times'] = click_log['industry'].astype(np.int64)*10000000 + click_log['click_times']


# click_log['creative_id_times'] = click_log['creative_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
# click_log['ad_id_times'] = click_log['ad_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
# click_log['product_id_times'] = click_log['product_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
# click_log['advertiser_id_times'] = click_log['advertiser_id'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
# click_log['product_category_times'] = click_log['product_category'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
# click_log['industry_times'] = click_log['industry'].astype(str).str.cat(click_log['click_times'].astype(str), sep='-')
click_log.head()

click_log = multi_column_LabelEncoder(click_log, 
                                      columns=['creative_id_times','ad_id_times','product_id_times',
                                               'advertiser_id_times','product_category_times', 'industry_times',
                                              ])

print(click_log['creative_id'].nunique())
print(click_log['ad_id'].nunique())
print(click_log['product_id'].nunique())
print(click_log['product_category'].nunique())
print(click_log['advertiser_id'].nunique())
print(click_log['industry'].nunique())
print(click_log['creative_id_times'].nunique())
print(click_log['ad_id_times'].nunique())
print(click_log['product_id_times'].nunique())
print(click_log['advertiser_id_times'].nunique())
print(click_log['product_category_times'].nunique())
print(click_log['industry_times'].nunique())

from collections import defaultdict
import sys
creative_id = defaultdict(list)
ad_id = defaultdict(list)
product_id = defaultdict(list)
product_category = defaultdict(list)
advertiser_id = defaultdict(list)
industry = defaultdict(list)
click_times = defaultdict(list)
time = defaultdict(list)
creative_id_times = defaultdict(list)
ad_id_times = defaultdict(list)
product_id_times = defaultdict(list)
advertiser_id_times = defaultdict(list)
product_category_times = defaultdict(list)
industry_times = defaultdict(list)


for user,f1,f2,f3,f4,f5,f6,f7,f8,f9,f10,f11,f12,f13,f14  in zip(click_log['user_id'], 
                            click_log['creative_id'],
                            click_log['ad_id'],
                            click_log['product_id'],
                            click_log['product_category'],
                            click_log['advertiser_id'],
                            click_log['industry'],
                            click_log['click_times'],
                            click_log['time'],
                            click_log['creative_id_times'],
                            click_log['ad_id_times'],
                            click_log['product_id_times'],
                            click_log['advertiser_id_times'],
                            click_log['product_category_times'],
                            click_log['industry_times']):
    creative_id[user].append(f1)
    ad_id[user].append(f2)
    product_id[user].append(f3)
    product_category[user].append(f4)
    advertiser_id[user].append(f5)
    industry[user].append(f6)
    click_times[user].append(f7)
    time[user].append(f8)
    
    creative_id_times[user].append(f9)
    ad_id_times[user].append(f10)
    product_id_times[user].append(f11)
    advertiser_id_times[user].append(f12)
    product_category_times[user].append(f13)
    industry_times[user].append(f14)
    
user_ids = pd.DataFrame({'user_id':list(creative_id.keys()),
                         'creative_id':list(creative_id.values()),
                        'ad_id':list(ad_id.values()),
                         'product_id':list(product_id.values()),
                         'product_category':list(product_category.values()),
                         'advertiser_id':list(advertiser_id.values()),
                         'industry':list(industry.values()),
                         'click_times':list(click_times.values()),
                         'time':list(time.values()),
                         
                         'creative_id_times':list(creative_id_times.values()),
                         'ad_id_times':list(ad_id_times.values()),
                         'product_id_times':list(product_id_times.values()),
                         'advertiser_id_times':list(advertiser_id_times.values()),
                         'product_category_times':list(product_category_times.values()),
                         'industry_times':list(industry_times.values()),
                        
                        
                        })
print(user_ids.shape)
user_ids.head()

del creative_id, ad_id, product_id, product_category, advertiser_id, industry,click_times,time, creative_id_times,
del ad_id_times,product_id_times,advertiser_id_times, product_category_times, industry_times
gc.collect()

save_pickle(user_ids, path_build + "user_ids.pkl")
