==> notebook2/1_data_nmt.ipynb <==
import sys
import os
import gc


from utils import *
import pandas as pd
import numpy as np

path_build = "./build/"
path_embed = "./embed/"
path_list = "./feature_series/"
path_mnt = "./mnt/"
#path_mnt = "./mnt_train/"
#path_mnt = "./mnt_train_mean/"

#cos_to_local("/home/tione/notebook/build/click_log.pkl", "lht-build-1259079995", "click_log.pkl")
user = pd.read_pickle(path_build + "train_user.pkl")
click_log = pd.read_pickle(path_build + "click_log.pkl")
train_click_log = click_log[click_log['user_id']<=3000000]

ctr_df = train_click_log.merge(user, how='left', on='user_id')
ctr_df = pd.concat([ctr_df, pd.get_dummies(ctr_df['age'],prefix='age'), pd.get_dummies(ctr_df['gender'],prefix='gender')], axis=1)
del ctr_df['gender_2']

gender_p = dict(user['gender'].value_counts() / 3000000)
age_p = dict(user['age'].value_counts() / 3000000)
print(gender_p, age_p)

ctr_list = ['gender_1', 'age_1', 'age_2', 'age_3', 'age_4', 'age_5', 'age_6', 'age_7', 'age_8', 'age_9', 'age_10']
del user
gc.collect()

numeric_features = ['product_id','product_category', 'advertiser_id', 'industry']
#numeric_features = [ 'ad_id', 'product_id', 'product_category', 'advertiser_id', 'industry']
mnt_df = pd.DataFrame()
mnt_df['user_id'] = range(1,4000001)
for fea in numeric_features:
    print(fea)
    creative_id_ctr = ctr_df.drop_duplicates(['user_id',fea])

    creative_id_ctr_df = creative_id_ctr.groupby([fea])[ctr_list].mean().add_prefix(f'prob_{fea}_').reset_index()
    creative_id_sum = creative_id_ctr.groupby([fea])[ctr_list].sum().add_prefix(f'sum_{fea}_').reset_index()
    creative_id_size = creative_id_ctr.groupby([fea]).size().reset_index().rename({0:'size'}, axis=1)
    creative_id_ctr_df[f'size'] = creative_id_size['size']
    #贝叶斯平滑
    for f in ctr_list:
        M = creative_id_ctr_df[f'prob_{fea}_{f}'].mean()
        S = creative_id_ctr_df[f'prob_{fea}_{f}'].var()
        alpha =( M * ((M*(1-M))/(S)-1)  )*2
        beta  =( (1-M) * ((M*(1-M))/(S)-1) )*2
        creative_id_ctr_df[f'sum_{fea}_{f}'] = creative_id_sum[f'sum_{fea}_{f}']
        creative_id_ctr_df[f'prob_{fea}_{f}'] = (creative_id_ctr_df[f'sum_{fea}_{f}'] + alpha) / (creative_id_ctr_df[f'size'] + alpha + beta)


    click_log_f = click_log.merge(creative_id_ctr_df, how='left', on=fea)

    #填充缺失值
    click_log_f[f'prob_{fea}_gender_1'] = click_log_f[f'prob_{fea}_gender_1'].fillna(gender_p[1])
    for i in range(1,11):
        click_log_f[f'prob_{fea}_age_{i}'] = click_log_f[f'prob_{fea}_age_{i}'].fillna(age_p[i])

    creative_id = pd.concat([
        click_log_f.groupby(['user_id'])[ [f'prob_{fea}_'+i for i in ctr_list] ].mean().add_prefix(f'mean_').reset_index(),
        click_log_f.groupby(['user_id'])[ [f'prob_{fea}_'+i for i in ctr_list] ].max().add_prefix(f'max_').reset_index().drop(['user_id'], axis=1),
        click_log_f.groupby(['user_id'])[ [f'prob_{fea}_'+i for i in ctr_list] ].min().add_prefix(f'min_').reset_index().drop(['user_id'], axis=1),
        click_log_f.groupby(['user_id'])[ [f'prob_{fea}_'+i for i in ctr_list] ].var().add_prefix(f'var_').reset_index().drop(['user_id'], axis=1)  
    ],axis=1)
    
    mnt_df = pd.concat([mnt_df, creative_id.drop(['user_id'], axis=1)], axis=1)

print(mnt_df.columns)
mnt_df.head()

#mnt_df.to_pickle(f"{path_mnt}/user_mnts.pkl")
mnt_df.columns

reduce_mem_usage(mnt_df)


mnt_df.to_pickle(f"{path_mnt}/user_mnts.pkl")
