==> notebook2/2_w2v (1).ipynb <==
!cat /proc/cpuinfo
!free -m

# 环境配置
!pip install gensim
!pip install Cython

import pandas as pd
import numpy as np
from gensim import models

import sys
#import Cython
from utils import *
# 导入数据
#cos_to_local("/home/tione/notebook/build/user_ids.pkl", "lht-build-1259079995", "user_ids.pkl")
#cos_to_local("/home/tione/notebook/build/train_user.pkl", "lht-build-1259079995", "train_user.pkl")

path_build = "./build/"
path_embed = "./embed/"
path_list = "./feature_series/"

user_ids = pd.read_pickle(f"{path_build}/user_ids.pkl")
user_ids.head()

w2v_features0 = [
    #{'name':'creative_id', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'ad_id', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    {'name':'advertiser_id', 'size':64, 'windows':10, 'min_count':1, 'version':1},
    {'name':'product_id', 'size':64, 'windows':10, 'min_count':1, 'version':1},
    {'name':'industry', 'size':32, 'windows':10, 'min_count':1, 'version':1},
    {'name':'product_category', 'size':16, 'windows':10, 'min_count':1, 'version':1},
    {'name':'time', 'size':16, 'windows':10, 'min_count':1, 'version':1},
    {'name':'click_times', 'size':8, 'windows':10, 'min_count':1, 'version':1},
    
    #{'name':'creative_id_times', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'ad_id_times', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'product_id_times', 'size':64, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'advertiser_id_times', 'size':64, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'product_category_times', 'size':32, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'industry_times', 'size':32, 'windows':10, 'min_count':1, 'version':1},
    
    
#     {'name':'time_clicktimes', 'size':91, 'windows':10, 'min_count':1, 'version':2,},
#     {'name':'time_creativeids', 'size':91, 'windows':10, 'min_count':1, 'version':2 },
    
]
w2v_features1 = [
    #{'name':'creative_id', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'ad_id', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'advertiser_id', 'size':128, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'product_id', 'size':128, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'industry', 'size':128, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'product_category', 'size':128, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'time', 'size':128, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'click_times', 'size':128, 'windows':10, 'min_count':1, 'version':1},
    
    #{'name':'creative_id_times', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'ad_id_times', 'size':256, 'windows':10, 'min_count':1, 'version':1},
#     {'name':'product_id_times', 'size':128, 'windows':10, 'min_count':1, 'version':1},
#     {'name':'advertiser_id_times', 'size':128, 'windows':10, 'min_count':1, 'version':1},
#     {'name':'product_category_times', 'size':128, 'windows':10, 'min_count':1, 'version':1},
#     {'name':'industry_times', 'size':128, 'windows':10, 'min_count':1, 'version':1},
    
    
    {'name':'product_advertiser', 'size':128, 'windows':10, 'min_count':1, 'version':1},
    {'name':'product_industry', 'size':128, 'windows':10, 'min_count':1, 'version':1},
    {'name':'advertiser_industry', 'size':128, 'windows':10, 'min_count':1, 'version':1},
    
#     {'name':'time_clicktimes', 'size':91, 'windows':10, 'min_count':1, 'version':2,},
#     {'name':'time_creativeids', 'size':91, 'windows':10, 'min_count':1, 'version':2 },
    
]
w2v_features2 = [
    #{'name':'creative_id', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'ad_id', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    {'name':'advertiser_id', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    {'name':'product_id', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    {'name':'industry', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    {'name':'product_category', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    {'name':'time', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    {'name':'click_times', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    
    #{'name':'creative_id_times', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'ad_id_times', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'product_id_times', 'size':64, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'advertiser_id_times', 'size':64, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'product_category_times', 'size':32, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'industry_times', 'size':32, 'windows':10, 'min_count':1, 'version':1},
    
    
#     {'name':'time_clicktimes', 'size':91, 'windows':10, 'min_count':1, 'version':2,},
#     {'name':'time_creativeids', 'size':91, 'windows':10, 'min_count':1, 'version':2 },
    
]

w2v_features = [
#     {'name':'creative_id', 'size':512, 'windows':10, 'min_count':1, 'version':1},
#     {'name':'ad_id', 'size':512, 'windows':10, 'min_count':1, 'version':1},
#     {'name':'advertiser_id', 'size':512, 'windows':10, 'min_count':1, 'version':1},
#     {'name':'product_id', 'size':512, 'windows':10, 'min_count':1, 'version':1},
#     {'name':'industry', 'size':512, 'windows':10, 'min_count':1, 'version':1},
#     {'name':'product_category', 'size':512, 'windows':10, 'min_count':1, 'version':1},
#     {'name':'time', 'size':512, 'windows':10, 'min_count':1, 'version':1},
#     {'name':'click_times', 'size':512, 'windows':10, 'min_count':1, 'version':1},
    
    #{'name':'creative_id_times', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'ad_id_times', 'size':256, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'product_id_times', 'size':64, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'advertiser_id_times', 'size':64, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'product_category_times', 'size':32, 'windows':10, 'min_count':1, 'version':1},
    #{'name':'industry_times', 'size':32, 'windows':10, 'min_count':1, 'version':1},
    
    
#     {'name':'time_clicktimes', 'size':91, 'windows':10, 'min_count':1, 'version':2,},
#     {'name':'time_creativeids', 'size':91, 'windows':10, 'min_count':1, 'version':2 },
    
]

for fea_dict in w2v_features1:
    print(fea_dict)
    name = fea_dict['name']
    size = fea_dict['size']
    windows = fea_dict['windows']
    min_count = fea_dict['min_count']
    version = fea_dict['version']
    print('start......')
    #np.save(f"{path_list}{name}_list_int.npy", user_ids[name])
    #user_ids[name] = user_ids[name].map(lambda x: [str(i) for i in x])
    #np.save(f"{path_list}{name}_list.npy", user_ids[name])
    feature_list = np.load(f"{path_list}{name}_list.npy", allow_pickle=True)
    w2v = models.Word2Vec(feature_list, sg=1, size=size, window=windows, workers=30, hs=1, min_count=min_count)
    w2v.wv.save_word2vec_format(f"{path_embed}/{name}_w2v_{windows}_{version}.bin",binary=True)
    #保存矩阵
    w2v = models.KeyedVectors.load_word2vec_format(f"{path_embed}/{name}_w2v_{windows}_{version}.bin",binary=True)
    if 'vocab_size' in fea_dict:
        w2v_vocab_size = fea_dict['vocab_size']
    else:
        w2v_vocab_size = len(w2v.vocab) + 5
        print(w2v_vocab_size)
    w2v_matrix = np.zeros((w2v_vocab_size, w2v.vector_size))
    for w in w2v.vocab:
        w2v_matrix[int(w)] = w2v[w]
    save_pickle(w2v_matrix.astype(np.float16), f"{path_embed}/{name}_w2v_matrix1.pkl",)
